fix: address review feedback on PR #1508
- Move pool config tests to src/config_tests.rs (separate file per CLAUDE.md convention)
- Update load_for_repo and save_for_repo debug logs to include profile_pool_state count
- Remove redundant top-level HashMap import and _use_hashmap dummy from state_tests.rs
- Add Auth Profile Pool section to docs/architecture.md documenting ProfileState and LRU selection strategy
